% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_model}
\alias{build_model}
\title{build_model}
\usage{
build_model(units, activation, input_shape, loss, optimizer)
}
\arguments{
\item{units}{vector list with number of neurons per layer (last layer has to be 1)}

\item{activation}{Activation function (either native from keras or custom)}

\item{input_shape}{number of neurons in the input layer, read from the pre processed dataset}

\item{loss}{loss function name (defaul: mean_absolute_error)}

\item{optimizer}{Keras optimizer}
}
\value{
Untrained model
}
\description{
define and compile keras model
}
\author{
Marcos Alves
}
