% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_baselines.R
\name{create_baselines}
\alias{create_baselines}
\title{create_baselines}
\usage{
create_baselines(modelid, targetdir, pprosdf)
}
\arguments{
\item{modelid}{model id output from \link{write_run_info}}

\item{targetdir}{output directory}

\item{pprosdf}{dataset}
}
\description{
run_analysis for cross valiation batches
}
\author{
Marcos Alves \email{mppalves@gmail.com}
}
