% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export2gams.R
\name{export2gams}
\alias{export2gams}
\alias{list_col_names}
\title{Export neural nets weights and biases to .csv and write gams code}
\usage{
export2gams(model, module, means, stddevs, min, max, inputs_vec, type, model_hash, flag)
}
\arguments{
\item{model}{trained model}

\item{module}{magpie module number and name}

\item{means}{named vector with feature means}

\item{stddevs}{named vetor with feature std}

\item{min}{min  scaling parameter}

\item{max}{max  scaling parameter}

\item{model_hash}{model id hash}

\item{flag}{flag used to identify the label column}

\item{inputs_vec}{vector with input column names}

\item{type}{Letter used to identify the dataset being used}
}
\description{
Save and export model weights and biases for use in GAMS
  reconstruction of neural networks.
}
\author{
Marcos Alves \email{mppalves@gmail.com}
}
