% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gym.R
\name{model_gym}
\alias{model_gym}
\title{model_gym}
\usage{
model_gym(pprosdf, ppsample, runmeta, modelid, features)
}
\arguments{
\item{pprosdf}{pre processed dataset}

\item{ppsample}{pre processed training dataset \link{ppsample}}

\item{runmeta}{list with compiled model and parameters.}

\item{modelid}{model id output from \link{write_run_info}}

\item{features}{boolean vector}
}
\value{
Trained model
}
\description{
Run model `fit` on the pre-processed sampled training data and perform model quality evaluations
}
\author{
Marcos Alves
}
