% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppsample.R
\name{ppsample}
\alias{ppsample}
\title{ppsample}
\usage{
ppsample(
  pprosdf,
  skip_timestep = NULL,
  p = 0.9,
  features = select_features(pprosdf),
  offset = 0
)
}
\arguments{
\item{pprosdf}{pre processed dataset}

\item{skip_timestep}{time step to be skipped in the sample (cross validation)}

\item{p}{train/test data split}

\item{features}{Boolean vector list containing the selected features}

\item{offset}{offset value for 0 centered scale outputs}
}
\value{
Sample training in matrix format
}
\description{
Processes the pre processed data set by selecting features, randomizing, scaling and slicing.
}
\author{
Marcos Alves
}
