% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{run_analysis}
\usage{
run_analysis(output_data, input_data, title, cor, unit, comment = NULL)
}
\arguments{
\item{output_data}{Model predictions}

\item{input_data}{Original data}

\item{title}{String used to name the plots and the exported files}

\item{cor}{Hexadecimal or color name used to color details in plots}

\item{unit}{unit measure for the output}

\item{comment}{Comment to printed in the .txt statistical file results}
}
\description{
Run Analysis take as input two sets of values (input and output
 data) and run a series of analysis outputing statistics and graphs to check
 how well the two datasets fit each other.
}
\author{
Marcos Alves \email{mppalves@gmail.com}
}
