% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_batch_analysis.R
\name{run_batch_analysis}
\alias{run_batch_analysis}
\title{run_batch_analysis}
\usage{
run_batch_analysis(
  modelid,
  targetdir,
  .cor = "green",
  .title = "test",
  .unit = "z"
)
}
\arguments{
\item{modelid}{model id output from \link{write_run_info}}

\item{targetdir}{output directory}

\item{.cor}{Hexadecimal or color name used to color details in plots}

\item{.title}{String used to name the plots and the exported files}

\item{.unit}{unit measure for the output}
}
\description{
run_analysis for cross valiation batches
}
\author{
Marcos Alves \email{mppalves@gmail.com}
}
