% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_batch_rmse.R
\name{run_batch_rmse}
\alias{run_batch_rmse}
\title{run_batch_rmse}
\usage{
run_batch_rmse(
  pprosdf,
  trained_model,
  features = select_features(pprosdf),
  targetdir,
  modelid,
  offset = 0,
  tag
)
}
\arguments{
\item{pprosdf}{pre processed dataset}

\item{trained_model}{trained keras model}

\item{features}{columns selected for training}

\item{targetdir}{output directory}

\item{modelid}{model id output from \link{write_run_info}}

\item{offset}{use if scaled output values are offset}

\item{tag}{flaged output}
}
\description{
run_analysis for cross valiation batches
}
\author{
Marcos Alves \email{mppalves@gmail.com}
}
