% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_move_files.R
\name{run_move_files}
\alias{run_move_files}
\title{run_batch_rmse}
\usage{
run_move_files(source_folder, modelid, targetdir, subtype)
}
\arguments{
\item{source_folder}{madrat source folder}

\item{modelid}{model id output from \link{write_run_info}}

\item{targetdir}{output directory}

\item{subtype}{dataset subtype}
}
\description{
run_analysis for cross valiation batches
}
\author{
Marcos Alves \email{mppalves@gmail.com}
}
