% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_features.R
\name{select_features}
\alias{select_features}
\title{select_features}
\usage{
select_features(pprosdf)
}
\arguments{
\item{pprosdf}{pre processed dataframe}
}
\value{
list containing `input_shape` and boolean vector of selected columns
}
\description{
Tool to select the relevant features used for training from the datasets
}
\author{
Marcos Alves
}
