% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_run_info.R
\name{write_run_info}
\alias{write_run_info}
\title{write_run_info}
\usage{
write_run_info(targetdir, run_title = "Soilc", runmeta, pprosdf)
}
\arguments{
\item{targetdir}{output directory}

\item{run_title}{run type name}

\item{runmeta}{list with compiled model and parameters.}

\item{pprosdf}{pre processed dataframe}
}
\value{
model hash id (character string)
}
\description{
Write compiled model info and generates model hash id
}
\author{
Marcos Alves
}
