% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprosdf.R
\name{pprosdf}
\alias{pprosdf}
\title{pprosdf}
\usage{
pprosdf(input, targetdir, repositories, flag)
}
\value{
Data frame
}
\description{
Pre processes the machine learning emulator based on Mapie input files
}
\author{
Marcos Alves
}
