% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsample.R
\name{tsample}
\alias{tsample}
\title{tsample}
\usage{
tsample(
  pprosdf,
  features,
  plot_test = F,
  tag,
  means_col,
  std_col,
  mean_lab,
  std_lab
)
}
\arguments{
\item{pprosdf}{input file}

\item{features}{output directory}

\item{plot_test}{Whether the output should reduced to cluster cells for plotting}

\item{tag}{flag}

\item{means_col}{means}

\item{std_col}{standard deviation}

\item{mean_lab}{means}

\item{std_lab}{standard deviation}
}
\value{
matrix
}
\description{
Pre processes validation dataset
}
\author{
Marcos Alves
}
