% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsample.R
\name{tsample}
\alias{tsample}
\title{tsample}
\usage{
tsample(pprosdf, features, plot_test = F, tag)
}
\arguments{
\item{pprosdf}{input file}

\item{features}{output directory}

\item{plot_test}{Whether the output should reduced to cluster cells for plotting}

\item{tag}{flag}
}
\value{
matrix
}
\description{
Pre processes validation dataset
}
\author{
Marcos Alves
}
